+++ CHANGELOG.md
@@ -81,6 +81,9 @@
 - sctbench/concurrent-software-benchmarks/micro_2_ok.c: add `pthread_join()`
 - sctbench/concurrent-software-benchmarks/micro_3_ok.c: add `pthread_join()`
 - sctbench/concurrent-software-benchmarks/token_ring_bad.c: add `pthread_join()`
+- sctbench/inspect_benchmarks/bzip2smp.comb.c: correct stdin and stdout redirection
+- sctbench/inspect_benchmarks/qsort_mt.c: thread routines return 0
+- sctbench/inspect_benchmarks/qsort.comb.c: add support for macOS
 
 ## Added
 - build action for macOS